Reorganize Intake files
