+++ Group2/src/main/java/com/stuypulse/robot/subsystems/Intake.java
@@ -1,4 +1,4 @@
-package com.stuypulse.robot.subsystems.intake;
+package com.stuypulse.robot.subsystems;
 
 import com.stuypulse.robot.constants.Settings;
 import com.stuypulse.robot.constants.Settings.Intake.Control;
@@ -40,7 +40,7 @@ public abstract class Intake extends SubsystemBase {
 
     /*** Drive Motors ***/
 
-    protected abstract void set(double speed);
+    public abstract void set(double speed);
 
     public void acquire() {
         set(Settings.Intake.ACQUIRE_SPEED.get());
@@ -50,11 +50,11 @@ public abstract class Intake extends SubsystemBase {
         set(Settings.Intake.DEACQUIRE_SPEED.get());
     }
     
-    protected abstract void stop();
+    public abstract void stop();
 
     /*** Deploy Motor ***/
 
-    protected abstract void setDeploy(double speed);
+    public abstract void setDeploy(double speed);
 
     public abstract Angle getAngle();
